requirements.txt
django
djangorestframework
requests
python-dotenv
google-generativeai
spotipy